==> notebooks/04_indicador.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

```{r,message=FALSE,warning=FALSE}
library(here)
library(lubridate)
library(outliers)
library(forecast)
library(Metrics)
library(tidyverse)

```

Rutas

```{r}
raw_data <- here("data", "raw")
processed_data <- here("data", "processed")

```

Cargando datos

```{r}
componentes = readRDS(paste0(processed_data, "/componentes.Rds"))
datos = readRDS(paste0(processed_data, "/datos_tasa_EDA.Rds"))
datos_long = readRDS(paste0(processed_data, "/datos_tasa.Rds"))

```



### Indicador

```{r}

#suma ponderado
#usando la proporcion de variancia explicada por cada componente

componentes<-componentes %>% mutate(IR= 0.93*PC1 + 0.056*PC2 + 0.011*PC3)


#Estandarizar indicador final

componentes<-componentes %>% 
  mutate(IR_S=(IR-mean(IR))/sd(IR))

summary(componentes)

```



### Validación 


```{r}

componentes %>%
  ggplot(aes(x = fecha, y = IR_S)) +
  geom_line() +
  geom_line(y=0,col="red",linetype = "dashed") +
  geom_line(y=median(componentes$IR_S),col="blue",linetype = "dashed") +
  theme(plot.title = element_text(hjust = 0.5)) +
  ylab("ICR") +
  xlab("Fecha") +
  ggtitle("Indicador")


```

```{r}

#promedio y mediana de cada semana para compararlo

datos_long$promedio<-apply(datos_long[,c(7:3606)],1,mean)
datos_long$mediana<-apply(datos_long[,c(7:3606)],1,median)

base_temp<-datos_long %>% select(semana_rango,fecha,promedio,mediana)


base_temp %>%
  pivot_longer( 
    cols = -c(semana_rango,fecha), # Toma todas menos id
    names_to = "Medida",
    values_to = "Valor"
  ) %>% 
  ggplot(aes(x = fecha, y = Valor,group=Medida,colour=Medida)) +
  geom_line() +
  theme(plot.title = element_text(hjust = 0.5)) +
  ylab("") +
  xlab("Fecha") +
  ggtitle("Promedio y mediana por semana")


```


```{r}
base_temp %>%
  inner_join(componentes %>% 
               select(fecha, IR_S), by = 'fecha') %>% 
  mutate(mediana = (mediana-mean(mediana))/sd(mediana)) %>% 
  pivot_longer( 
    cols = -c(semana_rango,fecha, promedio), # Toma todas menos id
    names_to = "Medida",
    values_to = "Valor"
  ) %>%
  ggplot(aes(x = fecha, y = Valor,group=Medida,colour=Medida)) +
  geom_line() +
  theme(plot.title = element_text(hjust = 0.5)) +
  ylab("") +
  xlab("Fecha") +
  ggtitle("Indicador y mediana estandarizada por semana")

```


```{r}
base_temp2 <- base_temp %>%
  inner_join(componentes %>% 
               select(fecha, IR_S), by = 'fecha') %>% 
  mutate(mediana = (mediana-mean(mediana))/sd(mediana)) 

tibble(Medida = c("MAPE", "MedAPE", "Sesgo %"),
       `Valor (%)` = c(
         mape(base_temp2$mediana, base_temp2$IR_S)*100,
         median(ape(base_temp2$mediana, base_temp2$IR_S))*100,
         percent_bias(base_temp2$mediana, base_temp2$IR_S)*100
       ))

```


La forma del indicador tiende a ser similar a la del promedio de los rendimientos en cada semana.

Niveles del ICR superiores a cero señalan aumentos en los rendimientos mayores al promedio histórico,

El indicador comenzó a mostrar un nivel mayor al promedio histórico a finales de 2017 en la semana del 19-07-17 al 25-07-17, el indicador alcanzó su máximo absoluto en la semana del 13-03-19 al 19-03-19, alrededor de 2.3 desviaciones sobre la media y mostró niveles superiores a la media hasta la semana del 23-10-19 al 29-10-19


A inicios del 2020 a partir de la semana del 18-03-20 al 24-03-20 se observa una marcada tendencia positiva a pesar de presentar niveles por debajo de la media.



**Descriptivo del indicador con series**

```{r}
componentes<-componentes %>% arrange(fecha)

serie<-ts(componentes$IR_S, frequency = 50,start = c(2015,40))

par(mfrow=c(3,1))

plot(serie,ylab="ICR",xlab="")
abline(h=0,col="red")
grid()

plot(diff(serie),ylab="diff(ICR)",xlab="")
abline(h=0,col="red")
grid()

plot(diff(log(serie)))
abline(h=0,col="red")
grid()
```



- Valores Extremos

Se detecta un posible valor extremo obtenido en la semana del 13-03-19 al 19-03-19. Según la prueba de Grubbs se rechaza la hipótesis nula de que 2.34 no es un outlier, mientras que en la prueba de Ji-cuadrado también se rechaza la hipótesis nula. 


```{r}
outlier(serie, opposite = FALSE, logical = FALSE)
#se rechaza
grubbs.test(serie, type = 10, opposite = FALSE, two.sided = TRUE) # one outlier
#se rechaza
chisq.out.test(serie, opposite = FALSE)

```



- Descomposición de la serie
`
```{r}

decomp = decompose(serie, "multiplicative")
plot(decomp)
```


```{r}
options(scipen = 999)
seasonplot(serie,col=rainbow(12),year.labels=TRUE,main="")
#factores estacionales
plot(decomp$seasonal,main="",xlab="",ylab="Factores")
monthplot(serie,col="blue",main="")
```








**Validación de curvas**


Análisis del comportamiento del indicador y de los factores no observables 


- 2020

Caso 1: aumento en el indicador

A inicios del 2020 a finales de marzo y en abril es donde se muestra un aumento pronunciado en el indicador.
En cuanto al nivel se observa un desplazamiento hacia arriba de las curvas a mayor fecha y mayor indicador, sin embargo en el largo plazo a partir de 9 años aproximadamente se invierte.
La tendencia se observan curvas ascendentes, las curvas a corto plazo tienen menores rendimientos que los de las curvas a largo plazo que son mayores. En cuanto a la curvatura vemos que a finales del corto plazo se empieza a formar.


```{r}

datos %>%
  filter(fecha=="2020-03-24" | fecha=="2020-03-31"|fecha=="2020-04-07" | fecha=="2020-04-21"  ) %>% 
  arrange(fecha) %>% 
  mutate(fecha=as.factor(fecha)) %>% 
  ggplot(aes(x = plazo_en_anos, y = tasa_percent, colour = fecha)) +
  geom_line() +
  geom_vline(xintercept=1,col="darkgray",linetype = "dashed")+
  geom_vline(xintercept=3,col="darkgray",linetype = "dashed")+
  theme(plot.title = element_text(hjust = 0.5)) +
  scale_x_continuous(breaks=seq(0, 10, 1))+
  ylim(0,15)+
  ylab("Tasas") +
  xlab("Plazo") +
  ggtitle("Curvas de rendimientos 2020")



```




Curvas en mayo y juni0 2020

```{r}

datos %>%
  filter((month(fecha)==5 | month(fecha)==6) & year(fecha)==2020) %>% 
  arrange(fecha) %>% 
  mutate(fecha=as.factor(fecha)) %>% 
  ggplot(aes(x = plazo_en_anos, y = tasa_percent, colour = fecha)) +
  geom_line() +
  geom_vline(xintercept=1,col="darkgray",linetype = "dashed")+
  geom_vline(xintercept=3,col="darkgray",linetype = "dashed")+
  theme(plot.title = element_text(hjust = 0.5)) +
  scale_x_continuous(breaks=seq(0, 10, 1))+
  ylim(0,15)+
  ylab("Tasas") +
  xlab("Plazo") +
  ggtitle("Curvas de rendimientos 2020")


```




- 2019

Caso 1: disminución en el indicador


Se observa un desplazamiento de la curva hacia abajo en corto, mediano y largo plazo. La curvatura se observa principalmente en el mediano plazo y  conforme el indicador es mayor, por ejemplo a inicios del periodo tomado, la pendiente de la curva se vuelve más pronunciada en el corto plazo. La tendencia es se observan curvas ascendentes.


```{r}

datos %>%
  filter(year(fecha)==2019 & (quarter(fecha)==2|quarter(fecha)==3|quarter(fecha)==4 ) ) %>% 
  arrange(fecha) %>% 
  mutate(fecha=as.factor(fecha)) %>% 
  ggplot(aes(x = plazo_en_anos, y = tasa_percent, colour = fecha)) +
  geom_line() +
  geom_vline(xintercept=1,col="darkgray",linetype = "dashed")+
  geom_vline(xintercept=3,col="darkgray",linetype = "dashed")+
  theme(
    plot.title = element_text(hjust = 0.5)) +
  scale_x_continuous(breaks=seq(0, 10, 1))+
  ylim(0,15)+
  ylab("Tasas") +
  xlab("Plazo") +
  ggtitle("Curvas de rendimientos 2019")

```



- 2017 : aumento en el indicador

En cuanto al nivel se observa un desplazamiento hacia arriba de las curvas a mayor fecha y mayor indicador.
La tendencia se observan curvas ascendentes, las curvas a corto plazo tienen menores rendimientos que los de las curvas a largo plazo que son mayores. La curvatura se observa principalmente en el mediano plazo.

```{r}

datos %>%
  filter(fecha=="2017-07-18"|fecha=="2017-07-25" ) %>% 
  arrange(fecha) %>% 
  mutate(fecha=as.factor(fecha)) %>% 
  ggplot(aes(x = plazo_en_anos, y = tasa_percent, colour = fecha)) +
  geom_line() +
  geom_vline(xintercept=1,col="darkgray",linetype = "dashed")+
  geom_vline(xintercept=3,col="darkgray",linetype = "dashed")+
  theme(
    plot.title = element_text(hjust = 0.5)) +
  scale_x_continuous(breaks=seq(0, 10, 1))+
  ylim(0,15)+
  ylab("Tasas") +
  xlab("Plazo") +
  ggtitle("Curvas de rendimientos 2017")

```





### Calibración






```{r}

library("fitdistrplus")

plotdist(componentes$IR, histo = TRUE, demp = TRUE)

descdist(componentes$IR, boot = 1000)

```

```{r}

plotdist(base_temp$mediana, histo = TRUE, demp = TRUE)

descdist(base_temp$mediana, boot = 1000)

```


```{r}
fb <- fitdist(base_temp$mediana, "weibull")
fg <- fitdist(base_temp$mediana, "gamma")
fln <- fitdist(base_temp$mediana, "lnorm")
par(mfrow = c(2, 2))
plot.legend <- c("Beta", "lognormal", "gamma")
denscomp(list(fb, fln, fg), legendtext = plot.legend)
qqcomp(list(fb, fln, fg), legendtext = plot.legend)
cdfcomp(list(fb, fln, fg), legendtext = plot.legend)
ppcomp(list(fb, fln, fg), legendtext = plot.legend)
```



```{r}
componentes %>%
  ggplot(aes(x = fecha, y = IR_S)) +
  geom_line() +
  geom_line(y=median(componentes$IR_S)-1,col="red",linetype = "dashed") +
  geom_line(y=median(componentes$IR_S)+1,col="red",linetype = "dashed") +
  geom_line(y=median(componentes$IR_S)-.5,col="orange",linetype = "dashed") +
  geom_line(y=median(componentes$IR_S)+.5,col="orange",linetype = "dashed") +
  geom_line(y=median(componentes$IR_S),col="blue",linetype = "dashed") +
  theme(legend.position = "none",
        plot.title = element_text(hjust = 0.5)) +
  ylab("ICR") +
  xlab("Fecha") +
  ggtitle("Indicador")

```



```{r}

refln = -0.0678186
sdln = 1.147298

#plotly::ggplotly(
componentes %>%
  ggplot(aes(x = fecha, y = IR_S)) +
  geom_line() +
  #geom_line(y=median(componentes$IR_S)-1,col="red",linetype = "dashed") +
  geom_line(y=refln+sdln,col="red",linetype = "dashed") +
  geom_line(y=refln-sdln,col="red",linetype = "dashed") +
  geom_line(y=refln+sdln*.75,col="orange",linetype = "dashed") +
  geom_line(y=refln-sdln*.75,col="orange",linetype = "dashed") +
  geom_line(y=refln ,col="green",linetype = "dashed") +
  #geom_line(y=median(componentes$IR_S),col="blue",linetype = "dashed") +
  theme(legend.position = "none",
        plot.title = element_text(hjust = 0.5)) +
  ylab("ICR") +
  xlab("Fecha") +
  ggtitle("")
#)
```

### Interpretación

Correlaciones por plazo
```{r}

curv_corto <- datos_long %>% 
  inner_join(base_temp2 %>% dplyr::select(fecha, IR_S), by = "fecha") %>% 
  dplyr::select(c(7:366, 3609)) 

curv_mediano <- datos_long %>% 
  inner_join(base_temp2 %>% dplyr::select(fecha, IR_S), by = "fecha") %>% 
  dplyr::select(c(367:1086), 3609)

curv_largo <- datos_long %>% 
  inner_join(base_temp2 %>% dplyr::select(fecha, IR_S), by = "fecha") %>% 
  dplyr::select(c(1087:3606), 3609)

cor_corto = cor(curv_corto)

ref_corto = which.max(cor_corto[1:360,361])

cor_mediano = cor(curv_mediano)

ref_mediano = which.max(cor_mediano[1:720,721])

cor_largo = cor(curv_largo)

ref_largo = which.max(cor_largo[1:2520,2521])

```

Medio año, 3 años y 5 años


```{r}

puntos_base = tibble(
  fecha = datos_long %>%
    inner_join(base_temp2 %>% dplyr::select(fecha, IR_S), by = "fecha") %>%
    dplyr::select(fecha) %>% 
    unlist() %>% 
    as_date(),
  indicador = datos_long %>%
    inner_join(base_temp2 %>% dplyr::select(fecha, IR_S), by = "fecha") %>%
    dplyr::select(IR_S) %>% 
    unlist(),
  curv_corto = unlist(curv_corto[ref_corto]),
  curv_mediano = unlist(curv_mediano[ref_mediano]),
  curv_largo = unlist(curv_largo[ref_largo])
) %>% 
  arrange(fecha)

puntos_base <- puntos_base %>% 
  mutate(corto_s = (curv_corto-mean(curv_corto))/sd(curv_corto),
         mediano_s = (curv_mediano-mean(curv_mediano))/sd(curv_mediano),
         largo_s = (curv_largo-mean(curv_largo))/sd(curv_largo)
         )

puntos_base <- puntos_base %>% 
  mutate(dif_corto = c(0,diff(curv_corto))*100,
         dif_mediano = c(0,diff(curv_mediano))*100,
         dif_largo = c(0,diff(curv_largo))*100,
         dif_corto_s = c(0,diff(corto_s)),
         dif_mediano_s = c(0,diff(mediano_s)),
         dif_largo_s = c(0,diff(largo_s)),
         diff_ind = c(0,diff(indicador))
         
         )

puntos_base <- puntos_base %>% 
  mutate(ind_pb_corto = round((dif_corto*diff_ind)/dif_corto_s,3), 
         ind_pb_mediano = round((dif_mediano*diff_ind)/dif_mediano_s,3),
         ind_pb_largo = round((dif_largo*diff_ind)/dif_largo_s,3)
         )


```

Pb corto plazo
```{r}
library(latticeExtra)

 
# --> construct separate plots for each series
obj1 <-
  xyplot(
    indicador ~ fecha,
    puntos_base,
    type = "l" ,
    lwd = 2,
    ylab = "",
    par.settings = list(superpose.line = list(
      col = c("black", "blue"),
      lwd = 2
    ))
  ) + layer(panel.ablineq(refln + sdln * .75, label = "", col = "orange"),
            panel.ablineq(refln + sdln, label = "", col = "red"),
            panel.ablineq(refln - sdln * .75, label = "", col = "orange"),
            panel.ablineq(refln - sdln, label = "", col = "red")
            )

obj2 <- xyplot(ind_pb_corto ~ fecha, puntos_base, type = "l", lwd=2, ylab = "")
 
# --> Make the plot with second y axis AND legend:
doubleYScale(obj1, obj2, text = c("Indicador", "Cambio en pb corto plazo") , add.ylab2 = TRUE )+
  latticeExtra::layer(panel.abline(v = 1))

```


Pb mediano plazo
```{r}
 
# --> construct separate plots for each series
obj1 <-
  xyplot(
    indicador ~ fecha,
    puntos_base,
    type = "l" ,
    lwd = 2,
    ylab = "",
    par.settings = list(superpose.line = list(
      col = c("black", "blue"),
      lwd = 2
    ))
  ) + layer(panel.ablineq(refln + sdln * .75, label = "", col = "orange"),
            panel.ablineq(refln + sdln, label = "", col = "red"),
            panel.ablineq(refln - sdln * .75, label = "", col = "orange"),
            panel.ablineq(refln - sdln, label = "", col = "red")
            )

obj2 <- xyplot(ind_pb_mediano ~ fecha, puntos_base, type = "l", lwd=2, ylab = "")
 
# --> Make the plot with second y axis AND legend:
doubleYScale(obj1, obj2, text = c("Indicador", "Cambio en pb mediano plazo") , add.ylab2 = TRUE )

```

Pb largo plazo
```{r}
 
# --> construct separate plots for each series

obj1 <-
  xyplot(
    indicador ~ fecha,
    puntos_base,
    type = "l" ,
    lwd = 2,
    ylab = "",
    par.settings = list(superpose.line = list(
      col = c("black", "blue"),
      lwd = 2
    ))
  ) + layer(panel.ablineq(refln + sdln * .75, label = "", col = "orange"),
            panel.ablineq(refln + sdln, label = "", col = "red"),
            panel.ablineq(refln - sdln * .75, label = "", col = "orange"),
            panel.ablineq(refln - sdln, label = "", col = "red")
            )

obj2 <- xyplot(ind_pb_largo ~ fecha, puntos_base, type = "l", lwd=2, ylab = "")
 
# --> Make the plot with second y axis AND legend:
doubleYScale(obj1, obj2, text = c("Indicador", "Cambio en pb largo plazo") , add.ylab2 = TRUE )

```
